detect device function

+++ django_server/DataManager.py
@@ -6,6 +6,9 @@ import serial
 import struct
 import binascii
 import time
+import os
+from .detect_device import serial_ports
+
 DB_URL = 'mongodb://localhost:27017'
 DB_NAME = 'PIVOTAL_DB'
 DB_COLLECTION = 'NodeData'
@@ -230,14 +233,22 @@ def getQueue():
             data = q.pop()
             sendData(data)
 
-ser = serial.Serial(port='/dev/cu.usbmodem14201', baudrate=9600, timeout=0.05)
+try:
+    print(serial_ports())
 
-# Start Threads
-t1 = threading.Thread(target=addData)
-t2 = threading.Thread(target=getQueue)
+    path = '/dev/cu.usbmodem14201'
+    if os.path.exists(path):
+        ser = serial.Serial(port=path, baudrate=9600, timeout=0.05)
+        # Start Threads
+        t1 = threading.Thread(target=addData)
+        t2 = threading.Thread(target=getQueue)
 
-time.sleep(2)
-deleteAll()
+        time.sleep(2)
+        deleteAll()
 
-t1.start()
-t2.start()
+        t1.start()
+        t2.start()
+    # else:
+    #     raise ValueError("No device")
+except Exception as e:
+    raise e
